Tighten up match for Contact content elements
Without this patch the behavior test for the contact is too loose and
the footer will match if the h1 element is removed.  This is a problem
because the test returns a false positive.

This patch addresses the problem by matching on the h1 element itself
with the expected text content.