Test ApplicationHelper#full_title mixin
Without this patch the full_title method that's mixed into controllers
is not explicitly tested.  This is a problem because we need to
re-implement this method in spec/support/utilities

This patch addresses the problem by explicitly adding coverage for the
helper method and updating the support utilities to simply include the
application helper class.