Add sign_in and current_user helpers (8.2.3)
Without this patch we don't have a way to track the currently signed in
user.  This patch addresses the problem by implementing two methods as
helpers for the Controllers and Views.  sign_in will compare a
remember_token permanently stored in the session with the hashed token
in the database and set current_user accordingly.

In addition, the current_user helper will return an instance of the User
model if there is a currently signed in user.