Customize the header for signed in users (8.2.4)
Without this patch the header is not dynamically updated after a user
signs in.  We need to give them a customized interface to see their
profile, settings, remove the sign in link, and sign out.

This patch addresses the problem by updating the layout header to behave
accordingly when a user is signed in.