Add association between User and Micropost (10.1.3)
Without this patch the User model is not associated with the Micropost
model.  This is a problem because we need to have access to the rails
idioms of using models through their association, e.g. user.microposts
and micropost.user

This patch addresses the problem by specifying a Micropost belongs to an
User and a User has many Microposts.