Return Microposts in descending order (10.1.4)
Without this patch the ORM returns Microposts in the order of their ID
number.  This is a problem because Microposts need to be returned in
descending order with the most recently created post at the beginning of
the Array.

This should fail because by default the posts will be ordered by id,
i.e., [older_micropost, newer_micropost]. This test also verifies the
basic correctness of the has_many association itself, by checking (as
indicated in Table 10.11) that user.microposts effectively returns an
array of microposts. The to_a method, discussed before in Section 4.3.1,
converts @user.microposts from its default state (which happens to be an
Active Record “collection proxy”) to a proper array appropriate for
comparison with the one we constructed by hand.

To get the ordering test to pass, we use a Rails facility called
default_scope with an order argument, as shown in Listing 10.11. (This
is our first example of the notion of scope. We will learn about scope
in a more general context in Chapter 11.)