Destroy microposts when user is destroyed (10.1.4)
Without this patch microposts associated with an user account persist in
the database when the user is destroyed.  This is a problem because the
microposts are not valid without an user object.

This patch addresses the problem by adding a test to make sure all posts
from a given user are destroyed when the user itself is destroyed.

The implementation uses the dependent: :destroy option to the `has_many`
relationship on the User model.