Add Micropost content validations (10.1.5)
Without this patch the system imposes no limitations on the length of
the content of a micropost.  This is a problem because blank posts and
long posts are invalid.

This patch addresses the problem by adding a maximum length of 140
characters for the content of each micropost and also ensures there is
some content.