Control access to Micropost#{create,destroy} (10.3.1)
Without this patch there is no access control on the create and destroy
actions for the MicropostsController and associated RESTful routes.
This is a problem because only signed in users should be able to create
and destroy Microposts.

With the code in Listing 10.24, the signed_in_user method is now
available in the Microposts controller, which means that we can restrict
access to the create and destroy actions with the before filter shown in
Listing 10.25. (Since we didn’t generate it at the command line, you
will have to create the Microposts controller file by hand.)