Add ability to post from the home page (10.3.2)
Without this patch a signed in user has no way to post a new micropost
from the web app itself.  This patch addresses the problem by adding a
new form for microposts directly to the home page of the app.

This patch also updates the error_messages shared partial to take a
passed in object rather than hard-coded against the @users instance
variable.