Add a proto-feed to the Home page (10.3.3)
Without this patch there is no post feed shown on the home page.  This
patch addresses the problem by creating two partials, shared/feed for
the feed itself and shared/feed_item for each item in the feed.

The feed is displayed on the home page and automatically updated when a
new post is made.  In the event of an error posting to the feed the feed
item list is set to an empty array.