Add relationships table tracking followers (11.1.1)
Without this patch the application has no way to track relationships
between users.  This is a problem because users need to be able to
follow other users and know who follows them.

This patch addresses the problem by adding a relationships table with
indexes suitable for a has_many through ActiveRecord relationship.