Add user/relationship associations (11.1.2)
Without this patch the User model is not associated with the
Relationship model in ActiveRecord.  This is a problem because users
needs to be able to follow other users and track who follows them.

This patch addresses the problem by associating User has many
relationships and Relationship belongs to follower and followed as class
User.