Followed users (11.1.4)
Without this patch there are no methods on the User model to follow and
unfollow other users.  This is a problem because the application needs
to have a clear and direct way to check the status of the relationship
between two users, follow another user, and unfollow another user.

This patch addresses the problem by implementing
User#{following?,follow!,unfollow!} instance methods on the User model.

In addition, this patch adds the following relationship; User has many
followed users through the relationships table with the source in the
followed_id column.